Move publishable and init-messages into simple `types` file.  Stop importing from `index`.
Also remove console logging.
