+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cc-sharing",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "communication between sharinator iframes ...",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
