+++ src/index.ts
@@ -1,7 +1,6 @@
 export * from "./iframe-phone";
-export * from "./sharing-client";
-export * from "./sharing-relay";
+export * from "./types";
 export * from "./sharing-parent";
-export * from "./publishable";
-export * from "./init-message";
+export * from "./sharing-client";
 export * from "./child-connection";
+export * from "./sharing-relay";
